#25 [FEAT] TeamViewModel, AdapterViewModel 로직 분리
