+++ presentation/src/main/java/com/tgyuu/presentation/feature/team/TeamViewModel.kt
@@ -23,9 +23,6 @@ import javax.inject.Inject
 
 @HiltViewModel
 class TeamViewModel @Inject constructor(
-    private val getMemberUseCase: GetMemberUseCase,
-    private val deleteMemberUseCase: DeleteMemberUseCase,
-    private val updateMemberInformationUseCase: UpdateMemberInformationUseCase,
     private val updateTeamInformationUseCase: UpdateTeamInformationUseCase,
     private val getTeamUseCase: GetTeamUseCase,
     @IO private val ioDispatcher: CoroutineDispatcher
@@ -46,55 +43,6 @@ class TeamViewModel @Inject constructor(
         object AddMember : TeamEvent()
         object ChangeTeamName : TeamEvent()
         object ChangeTeamImage : TeamEvent()
-    }
-
-    private val _memberListFlow = MutableStateFlow<UiState<List<Member>>>(UiState.Init)
-    val memberListFlow = _memberListFlow.asStateFlow()
-
-    private fun setMemberListState(uiState: UiState<List<Member>>) {
-        _memberListFlow.value = uiState
-    }
-
-    fun getMemberList() {
-        setMemberListState(UiState.Loading)
-        viewModelScope.launch(ioDispatcher) {
-            getMemberUseCase.getAllMembers().collect { setMemberListState(UiState.Success(it)) }
-        }
-    }
-
-    var updateMember: Member? = null
-
-    fun updateMemberImage(image: String) {
-        if (updateMember == null) {
-            return
-        }
-
-        viewModelScope.launch(ioDispatcher) {
-            updateMemberInformationUseCase.updateMemberImage(updateMember!!, image)
-        }
-        getMemberList()
-    }
-
-    fun removeMemberImage() {
-        if (updateMember == null) {
-            return
-        }
-
-        viewModelScope.launch(ioDispatcher) {
-            updateMemberInformationUseCase.removeMemberImage(updateMember!!)
-        }
-        getMemberList()
-    }
-
-    fun removeMember() {
-        if (updateMember == null) {
-            return
-        }
-
-        viewModelScope.launch(ioDispatcher) {
-            deleteMemberUseCase(updateMember!!)
-        }
-        getMemberList()
     }
 
     private val _team = MutableStateFlow<UiState<Team>>(UiState.Init)
